test(dedupe): consolidate perps_baseline tests (5 → 3 files)
Cluster bcf5f1feb7e9: merge similar_names test files for perps_baseline

Files consolidated:
- test_perps_baseline.py (NEW): merged init_and_guards + models (145 lines)
- test_perps_baseline_signals.py (NEW): merged entry_and_exit + indicators (214 lines)
- test_perps_baseline_strategies.py: renamed from spot_and_perps_strategies (104 lines)

Removed:
- test_perps_baseline_baseline_strategy_init_and_guards.py
- test_perps_baseline_entry_and_exit_signals.py
- test_perps_baseline_indicators_and_confidence.py
- test_perps_baseline_models.py

All 27 tests pass.
