+++ tests/unit/gpt_trader/features/live_trade/strategies/test_perps_baseline_strategies.py
